consolidated tests into single integration test file
